rm_node deletes paritions.
This is refactor of commit e0762a993.

Partiiton deletion implemented with trigger on row removal. We are in process of
switching replication model from sausage to spider, you know, and this code
assumes that there are no more than two partitions, in which case models are
equal.

Since we assume only 2 copies of data, after primary removal there is no need to
choose new primary and rebuilt logical channels from it to other copies. Because
of that, we implemented 'promotion' in trigger on part removal. Basically, we
just mark partition as primary, replace fdw -> usual, usual -> fdw on nodes with
replica and old primary and update fdw everywhere.