Review for 82f8d62a545, creating LR slots in separate xacts.
Without this, creating replicas for large number of partitions leads to
strange visibility bug: historic snapshot in walsender doesn't see
publication which was created in one xact with replication slot. This patch
helps, but are still not sure on the root of the problem, so we need to
investigate this more.

Besides, for more safety we divide here slots dropping & creation in separate
xacts, probably we will unite them back again later.