mv_partition failure cleanup: tmp LR channel and tables.
This is not yet tested properly, we need mv_partition to fail at arbitrary time
during load for that.

Also, fix LR channels first, only then FDW mappings in recover(). This ensures
that replication is configured appropriately when nodes learn about part
locations. Theoretically, for mv_partition we could arrange things so that it
doesn't matter; however, for correct failover (rm_node) this will be necessary.